Extract output schemas from mcp.tools
Second of three planned PRs decomposing backend/mcp/tools.py. This pass
extracts all per-tool output schema builders and intermediate
result-shape helpers — pure mechanical movement, no behavior change.

New file:
- tools/output_schemas.py (~970 LOC) — 47 schema builder functions:
  per-tool output_schema (_search_agreements_output_schema, etc.),
  result-shape helpers (_agreement_search_result_schema, etc.),
  tool-capabilities meta schemas (_tool_capabilities_output_schema,
  _server_capabilities_output_schema, _workflow_output_schema,
  _field_inventory_schema, _concept_note_schema, etc.)

The two _CAPABILITIES_SECTIONS_* tuples that sat in the middle of the
extracted range moved to tools/constants.py (they're shared between the
output schema and the runtime payload).

tools/__init__.py: 4454 -> 3539 LOC.
